+++ niffler-e-2-e-tests/build.gradle
@@ -3,7 +3,7 @@ buildscript {
         protobufVersion = '3.21.12'
         grpcVersion = '1.52.1'
         allureVersion = '2.21.0'
-        selenideVersion = '6.17.0'
+        selenideVersion = '6.13.0'
     }
 }
