--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/api/CategoryService.java
@@ -1,11 +1,0 @@
-package guru.qa.niffler.api;
-
-import guru.qa.niffler.model.CategoryJson;
-import retrofit2.Call;
-import retrofit2.http.Body;
-import retrofit2.http.POST;
-
-public interface CategoryService {
-    @POST("/category")
-    Call<CategoryJson> addCategory(@Body CategoryJson categoryJson);
-}
