+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/Spend.java
@@ -1,6 +1,6 @@
-package guru.qa.niffler.jupiter.annotation;
+package guru.qa.niffler.jupiter;
 
-import guru.qa.niffler.jupiter.extension.CategoryExtension;
+import guru.qa.niffler.model.CurrencyValues;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.lang.annotation.ElementType;
@@ -10,11 +10,17 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-@ExtendWith(CategoryExtension.class)
-public @interface Category {
+@ExtendWith(SpendExtension.class)
+public @interface Spend {
 
     String username();
 
+    String description();
 
     String category();
+
+    double amount();
+
+    CurrencyValues currency();
+
 }
