--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/Spend.java
@@ -1,28 +1,0 @@
-package guru.qa.niffler.jupiter.annotation;
-
-import guru.qa.niffler.jupiter.extension.SpendExtension;
-import guru.qa.niffler.jupiter.extension.SpendExtensionParameterResolver;
-import guru.qa.niffler.model.CurrencyValues;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-@ExtendWith({SpendExtension.class, SpendExtensionParameterResolver.class})
-public @interface Spend {
-
-    String username();
-
-    String description();
-
-    String category();
-
-    double amount();
-
-    CurrencyValues currency();
-
-}
