--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/AroundTestExtension.java
@@ -1,21 +1,0 @@
-package guru.qa.niffler.jupiter.extension;
-
-import org.junit.jupiter.api.extension.BeforeAllCallback;
-import org.junit.jupiter.api.extension.ExtensionContext;
-
-public interface AroundTestExtension extends BeforeAllCallback {
-
-    default void onStart(ExtensionContext context) {
-    }
-
-    default void onEnd() {
-    }
-
-    @Override
-    default void beforeAll(ExtensionContext context) throws Exception {
-        context.getRoot().getStore(ExtensionContext.Namespace.GLOBAL).getOrComputeIfAbsent(this.getClass(), k -> {
-            onStart(context);
-            return (ExtensionContext.Store.CloseableResource) this::onEnd;
-        });
-    }
-}
