--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/BrowserExtension.java
@@ -1,42 +1,0 @@
-package guru.qa.niffler.jupiter.extension;
-
-import com.codeborne.selenide.Configuration;
-import com.codeborne.selenide.Selenide;
-import com.codeborne.selenide.WebDriverRunner;
-import io.qameta.allure.Allure;
-import org.junit.jupiter.api.extension.AfterEachCallback;
-import org.junit.jupiter.api.extension.BeforeAllCallback;
-import org.junit.jupiter.api.extension.ExtensionContext;
-import org.junit.jupiter.api.extension.TestExecutionExceptionHandler;
-import org.openqa.selenium.OutputType;
-import org.openqa.selenium.TakesScreenshot;
-
-import java.io.ByteArrayInputStream;
-
-public class BrowserExtension implements BeforeAllCallback, AfterEachCallback, TestExecutionExceptionHandler {
-
-
-    @Override
-    public void beforeAll(ExtensionContext context) throws Exception {
-        Configuration.browser = "chrome";
-        Configuration.browserSize = "1980x1024";
-    }
-
-    @Override
-    public void afterEach(ExtensionContext context) throws Exception {
-        if (WebDriverRunner.hasWebDriverStarted()) {
-            Selenide.closeWindow();
-            Selenide.closeWebDriver();
-        }
-    }
-
-    @Override
-    public void handleTestExecutionException(ExtensionContext context, Throwable throwable) throws Throwable {
-        if (WebDriverRunner.hasWebDriverStarted()) {
-            Allure.addAttachment("Screenshoot on fail", new ByteArrayInputStream(
-                    ((TakesScreenshot) WebDriverRunner.getWebDriver()).getScreenshotAs(OutputType.BYTES)
-            ));
-        }
-        throw throwable;
-    }
-}
