--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/CategoryJson.java
@@ -1,39 +1,0 @@
-package guru.qa.niffler.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.util.UUID;
-
-public class CategoryJson {
-    @JsonProperty("id")
-    private UUID id;
-    @JsonProperty("category")
-    private String category;
-    @JsonProperty("username")
-    private String username;
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public String getCategory() {
-        return category;
-    }
-
-    public void setCategory(String category) {
-        this.category = category;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-}
