--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/FriendState.java
@@ -1,5 +1,0 @@
-package guru.qa.niffler.model;
-
-public enum FriendState {
-    INVITE_SENT, INVITE_RECEIVED, FRIEND
-}
