--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/UserJson.java
@@ -1,103 +1,0 @@
-package guru.qa.niffler.model;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import guru.qa.niffler.jupiter.annotation.User.UserType;
-
-import java.util.UUID;
-
-public class UserJson {
-    @JsonProperty("id")
-    private UUID id;
-    @JsonProperty("username")
-    private String username;
-    @JsonProperty("firstname")
-    private String firstname;
-    @JsonProperty("surname")
-    private String surname;
-    @JsonProperty("currency")
-    private CurrencyValues currency;
-    @JsonProperty("photo")
-    private String photo;
-    @JsonProperty("friendState")
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private FriendState friendState;
-
-    transient String password;
-    transient UserType userType;
-
-    public UserJson() {
-    }
-
-    public UserType getUserType() {
-        return userType;
-    }
-
-    public void setUserType(UserType userType) {
-        this.userType = userType;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getFirstname() {
-        return firstname;
-    }
-
-    public void setFirstname(String firstname) {
-        this.firstname = firstname;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
-    public CurrencyValues getCurrency() {
-        return currency;
-    }
-
-    public void setCurrency(CurrencyValues currency) {
-        this.currency = currency;
-    }
-
-    public String getPhoto() {
-        return photo;
-    }
-
-    public void setPhoto(String photo) {
-        this.photo = photo;
-    }
-
-    public FriendState getFriendState() {
-        return friendState;
-    }
-
-    public void setFriendState(FriendState friendState) {
-        this.friendState = friendState;
-    }
-}
