--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/BaseWebTest.java
@@ -1,8 +1,0 @@
-package guru.qa.niffler.test;
-
-import guru.qa.niffler.jupiter.annotation.WebTest;
-
-@WebTest
-public abstract class BaseWebTest {
-
-}
